+++ CHANGELOG.md
@@ -2,3 +2,6 @@
 --
 #### 2015-03-06: Steven Velozo
 * Initial Code Dump
+
+#### 2015-03-09: Steven Velozo
+* Test coverage, worked out kinks